显示刷新率dfx重构

+++ rosen/modules/render_service/core/drawable/dfx/rs_dirty_rects_dfx.h
@@ -62,7 +62,7 @@ private:
     const std::unique_ptr<RSRenderParams>& displayParams_;
 
     bool RefreshRateRotationProcess(RSPaintFilterCanvas& canvas,
-        ScreenRotation rotation, float translateWidth, float translateHeight);
+        ScreenRotation rotation, int translateWidth, int translateHeight);
     void DrawCurrentRefreshRate(RSPaintFilterCanvas& canvas);
     void DrawDirtyRectForDFX(RSPaintFilterCanvas& canvas, RectI dirtyRect,
         const Drawing::Color color, const RSPaintStyle fillType, int edgeWidth = 6) const;
